+++ R/cache.setup.R
@@ -17,6 +17,8 @@
 #' If any other string is used a folder will be created within either the
 #' [getwd()] directory or,
 #' if available, within the directory identified by [here::here()].
+#' If using an RStudio project, these will create the directory in the
+#' project directory.
 #' @param interactive
 #' Logical.
 #' `TRUE` indicates that confirmation will be required before a cache directory
@@ -94,7 +96,7 @@
 #' @export
 #'
 #' @examples
-#' \donttest{
+#' \dontrun{
 #'   # Setup a cache directory
 #'   cache.setup()
 #'   # Now code with check = TRUE or save_out = TRUE will work, e.g.,